Re-factor model_plotter for better integration with IPython
This commit implements the following improvements:

 (1) adds "mystic.model_plotter.model_plotter(cmdargs)" function

 (2) adds an additional parameter -p/--plot-filepath


1) mystic.model_plotter.model_plotter(cmdargs)

This convenience functions allows to directly invoke the plotter
providing either (a) a list of arguments or (b) a string representing the
arguments.

For example, from IPython, ones could plot using the following approaches:

%matplotlib inline
from mystic.model_plotter import model_plotter
model_plotter(cmdargs='mystic.models.zimmermann log.txt -b "-5:10:.1, -5:10:.1" -d -x 1')

or

%matplotlib inline
from mystic.model_plotter import model_plotter
model_plotter(cmdargs=['mystic.models.zimmermann', 'log.txt', '-b', '-5:10:.1, -5:10:.1', '-d', '-x', '1'])


2) additional parameter -p/--plot-filepath

This new argument allows to specify an image filename where the generated
plot should be saved.

For example:

mystic_model_plotter.py mystic.models.zimmermann -b "-5:10:.1, -5:10:.1" -d -x 1 -p output.png
